+++ src/com/nixinova/mineo/world/blocks/Block.java
@@ -9,6 +9,8 @@ public class Block {
 	public static final Block STONE = new Block("blocks/stone");
 	public static final Block DIRT = new Block("blocks/dirt");
 	public static final Block GRASS = new Block("blocks/grass");
+	public static final Block LOG = new Block("blocks/log");
+	public static final Block LEAF = new Block("blocks/leaf");
 
 	public static final Block DEBUG = new Block("debug");
 	public static final Block MISSING = new Block("missing_texture");