+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next
+- Added jagged terrain
+- Added trees
+- Added new blocks: log and leaf
+
 ## 0.1.4
 - New save file format (v2.0) using run-length encoding
   - Saves roughly 500x file size and makes saving and loading much faster